Combat Distance fixes
Fixed a bug in charge distance causing it to return the wrong value, numerous cascading problems were resolved. Several related systems were fixed in the process (grim trigger distances for lings and for scourge).  Combat is improved in test scenarioes, still does not lose to stone.

Caused the bot to check for long distance mining and building expos if it is LD mining.

+++ Utilities.cpp
@@ -1586,7 +1586,7 @@ int CUNYAIModule::getProperRange(const UnitType u_type, const Player owner) {
 
 int CUNYAIModule::getChargableDistance(const Unit & u, const Unit_Inventory & ei_loc)
 {
-    return (int)( u->getType() != UnitTypes::Zerg_Lurker * (int)CUNYAIModule::getProperSpeed(u) * (int)ei_loc.max_cooldown_ + CUNYAIModule::getProperRange(u) ); //lurkers have a proper speed of 0.
+    return (u->getType() != UnitTypes::Zerg_Lurker) * (int)CUNYAIModule::getProperSpeed(u) * (int)ei_loc.max_cooldown_ + CUNYAIModule::getProperRange(u) ; //lurkers have a proper speed of 0.
 }
